refactor: remove Activity dependency from AnkiServer()
which also allowed removing the ReviewerServer and moving the responsibilities back to the Reviewer
